+++ lib/src/rules/data/repository_impl_purity.dart
@@ -1,9 +1,11 @@
 import 'package:custom_lint_builder/custom_lint_builder.dart';
 
+import '../../config/y_lints_config.dart';
 import '../_shared/_architecture_rule.dart';
 
 class RepositoryImplPurity extends ArchitectureRule {
-  const RepositoryImplPurity() : super(locationCode: _location);
+  RepositoryImplPurity({YLintsConfig? config})
+      : super(locationCode: _location, config: config);
 
   static const _location = LintCode(
     name: 'repository_impl_location',
@@ -16,5 +18,5 @@ class RepositoryImplPurity extends ArchitectureRule {
 
   @override
   bool isAllowedPath(String filePath) =>
-      filePath.contains('/lib/data/repositories/');
+      filePath.contains(config.dataRepositories);
 }