+++ lib/src/rules/domain/domain_entity_purity.dart
@@ -2,10 +2,12 @@ import 'package:analyzer/dart/ast/ast.dart';
 import 'package:analyzer/error/listener.dart';
 import 'package:custom_lint_builder/custom_lint_builder.dart';
 
+import '../../config/y_lints_config.dart';
 import '../_shared/_architecture_rule.dart';
 
 class DomainEntityPurity extends ArchitectureRule {
-  const DomainEntityPurity() : super(locationCode: _location);
+  DomainEntityPurity({YLintsConfig? config})
+      : super(locationCode: _location, config: config);
 
   static const _location = LintCode(
     name: 'domain_entity_location',
@@ -24,7 +26,7 @@ class DomainEntityPurity extends ArchitectureRule {
 
   @override
   bool isAllowedPath(String filePath) =>
-      filePath.contains('/lib/domain/entities/');
+      filePath.contains(config.domainEntities);
 
   @override
   void checkClassStructure(ClassDeclaration node, ErrorReporter reporter) {