+++ lib/src/rules/domain/repository_purity.dart
@@ -1,9 +1,11 @@
 import 'package:custom_lint_builder/custom_lint_builder.dart';
 
+import '../../config/y_lints_config.dart';
 import '../_shared/_architecture_rule.dart';
 
 class RepositoryPurity extends ArchitectureRule {
-  const RepositoryPurity() : super(locationCode: _location);
+  RepositoryPurity({YLintsConfig? config})
+      : super(locationCode: _location, config: config);
 
   static const _location = LintCode(
     name: 'repository_location',
@@ -16,5 +18,5 @@ class RepositoryPurity extends ArchitectureRule {
 
   @override
   bool isAllowedPath(String filePath) =>
-      filePath.contains('/lib/domain/repositories/');
+      filePath.contains(config.domainRepositories);
 }